requirements.txt
appdirs==1.4.4
Appium-Python-Client==5.0.0
assertpy==1.1
attrs==25.3.0
baron==0.10.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
chromedriver==2.24.1
click==8.1.8
ConfigArgParse==1.7
Cython==3.0.12
debugpy==1.8.13
Faker==37.1.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
genson==1.3.0.1
get-chrome-driver==1.4
getgauge==0.4.8
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
idna==3.10
imgcompare==2.0.1
isim==15.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
locust==2.33.2
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
msgpack==1.1.0
numpy==2.2.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
protobuf==3.20.3
psutil==7.0.0
pure-python-adb==0.3.0.dev0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.1.0
pytweening==1.2.0
pyzmq==26.3.0
redbaron==0.9.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rply==0.7.8
rubicon-objc==0.5.0
screeninfo==0.8.1
selenium==4.31.0.202503270011
setuptools==78.1.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
toml
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2